backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "A FastAPI backend for data processing and analysis"
authors = ["Rocha,L."]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.115.12,<0.116.0"
uvicorn = ">=0.34.2,<0.35.0"
requests = ">=2.32.3,<3.0.0"
httpx = ">=0.28.1,<0.29.0"
python-dotenv = ">=1.1.0,<2.0.0"
polars = ">=1.29.0,<2.0.0"
duckdb = ">=1.2.2,<2.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
